Big update - stable no new "features"
Implemented new card storage system. Old cards are now display cards,
and contain a "card" object that is just information. Admin states are
no longer planned to be used. As far as I can tell, no new bugs are
introduced. To build the total card library, new cards will be put into
lua files in the cards directory and simply instantiate the proper card
object, then adding it to the master list. This is not shown yet, but
the framework is there.